+++ src/codegen/expressions/binary_expressions.rs
@@ -1,11 +1,9 @@
 use inkwell::{FloatPredicate, IntPredicate};
 use inkwell::builder::{Builder, BuilderError};
-use inkwell::types::BasicTypeEnum;
-use inkwell::values::{BasicValueEnum, IntValue, PointerValue};
+use inkwell::values::{BasicValueEnum, IntValue};
 
 use crate::codegen::Compiler;
-use crate::data::mlir::{MlirExpr, MlirExprKind};
-use crate::util::str_intern::InternedStr;
+use crate::data::mlir::MlirExpr;
 
 macro_rules! build_arithmetic_binop {
     (
@@ -65,33 +63,6 @@ macro_rules! build_arithmetic_binop {
     };
 }
 
-macro_rules! build_logical_binop {
-    ($compiler:ident, $left:ident, $left:ident, $build_method:ident) => {{
-        let (logical_left, logical_right) = $compiler.compile_binary_logical_expr($left, $right);
-
-        let value = $compiler
-            .builder()
-            .$build_method(logical_left, logical_right, stringify!($build_method))
-            .unwrap();
-
-        BasicValueEnum::from(value)
-    }};
-}
-
-macro_rules! build_bitwise_binop {
-    ($compiler:ident, $left:ident, $right:ident, $build_method:ident) => {{
-        let left = $compiler.compile_expression($left).into_int_value();
-        let right = $compiler.compile_expression($right).into_int_value();
-
-        let value = $compiler
-            .builder()
-            .$build_method(left, right, stringify!($build_method))
-            .unwrap();
-
-        BasicValueEnum::from(value)
-    }};
-}
-
 impl<'a, 'mlir, 'ctx> Compiler<'a, 'mlir, 'ctx> {
     #[inline(always)]
     fn compile_binary_expr(
@@ -103,62 +74,6 @@ impl<'a, 'mlir, 'ctx> Compiler<'a, 'mlir, 'ctx> {
             self.compile_expression(left),
             self.compile_expression(right),
         )
-    }
-
-    pub(super) fn compile_assignment(
-        &mut self,
-        left: &MlirExpr,
-        right: &MlirExpr,
-    ) -> BasicValueEnum<'ctx> {
-        debug_assert!(left.is_lval);
-        let assign_value = self.compile_expression(right);
-
-        let assign_ptr = match &*left.kind {
-            MlirExprKind::Variable(id) => self.get_pointer(id),
-            MlirExprKind::Deref(expr) => self.compile_expression(expr).into_pointer_value(),
-            MlirExprKind::Member(_struct, member) => self.get_struct_member_pointer(
-                self.convert_type(&left.ty.as_basic()),
-                _struct,
-                member,
-            ),
-            MlirExprKind::Index(array, index) => {
-                self.get_array_index_pointer(self.convert_type(&left.ty.as_basic()), array, index)
-            }
-            _ => self.compile_expression(left).into_pointer_value(),
-        };
-
-        self.builder().build_store(assign_ptr, assign_value);
-
-        assign_value
-    }
-
-    fn get_struct_member_pointer(
-        &mut self,
-        access_ty: BasicTypeEnum<'ctx>,
-        _struct: &MlirExpr,
-        member: &InternedStr,
-    ) -> PointerValue<'ctx> {
-        let struct_ident = _struct.ty.kind.get_struct_ident();
-        let _struct = self.compile_expression(_struct).into_pointer_value();
-        let member_offset = self.mlir.get_struct_member_offset(&struct_ident, member);
-        self.builder()
-            .build_struct_gep(access_ty, _struct, member_offset, "get_struct_member_ptr")
-            .unwrap()
-    }
-
-    fn get_array_index_pointer(
-        &mut self,
-        access_ty: BasicTypeEnum<'ctx>,
-        array: &MlirExpr,
-        index: &MlirExpr,
-    ) -> PointerValue<'ctx> {
-        let array_ptr = self.compile_expression(array).into_pointer_value();
-        let index_value = self.compile_expression(index).into_int_value();
-        unsafe {
-            self.builder()
-                .build_gep(access_ty, array_ptr, &[index_value], "get_array_index_ptr")
-                .unwrap()
-        }
     }
 
     pub(super) fn compile_addressof(&mut self, expr: &MlirExpr) -> BasicValueEnum<'ctx> {
@@ -385,8 +300,11 @@ impl<'a, 'mlir, 'ctx> Compiler<'a, 'mlir, 'ctx> {
         build_closure: BuildClosure,
     ) -> BasicValueEnum<'ctx>
     where
-        BuildClosure:
-            Fn(&Builder, IntValue<'ctx>, IntValue<'ctx>) -> Result<IntValue<'ctx>, BuilderError>,
+        BuildClosure: Fn(
+            &Builder<'ctx>,
+            IntValue<'ctx>,
+            IntValue<'ctx>,
+        ) -> Result<IntValue<'ctx>, BuilderError>,
     {
         let left = self.compile_expression(left).into_int_value();
         let right = self.compile_expression(right).into_int_value();
